requirements.txt
fastapi
uvicorn
requests
python-dotenv
pandas
fpdf
jinja2
pytest
